+++ lib/abstract_syntax/program.ml
@@ -22,6 +22,7 @@ type t =
   { f_rules : f_rules_by_name
   ; g_rules : g_rules_by_name
   ; extract_f_rules : Symbol_set.t
+  ; productive_g_rules : Symbol_set.t
   }
 
 let find_f_rule ~program op : f_rule =